BUG: Prevent unloading of plugins
Not unloading plugins is Qt's policy, as decided during the 5.0
development process and reaffirmed in 5.6. This is due to static
data in plugins leaking out and remaining in use past the unloading
of the plugin, causing crashes.

This happens for example with Qt Print Support on Windows and macOS,
which uses Q_IMPORT_PLUGIN() to register static plugins. Without
the appropriate hint, the same static plugin is registered for each
and every plugin that indirectly depends on Qt Print Support. The meta
data of these registered static plugins will be freed yet still
referenced in the global list of static plugins, leading to crashes.